two_duckies/requirements.txt
oauthlib==1.1.2
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
tweepy==3.5.0
picamera==1.12
PyYAML==3.12